% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts.R
\name{cl_lts_palette}
\alias{cl_lts_palette}
\title{Colours for Level of Traffic Stress}
\usage{
cl_lts_palette()
}
\value{
A named character vector of four colours, LTS 1 to 4.
}
\description{
Colours for Level of Traffic Stress
}
